Add ASTFormalComment
