+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/ast/AbstractApexNode.java
@@ -4,8 +4,6 @@
 
 package net.sourceforge.pmd.lang.apex.ast;
 
-import net.sourceforge.pmd.lang.ast.AbstractNode;
-import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.ast.SourceCodePositioner;
 
 import apex.jorje.data.Location;
@@ -13,7 +11,7 @@ import apex.jorje.data.Locations;
 import apex.jorje.semantic.ast.AstNode;
 import apex.jorje.semantic.exception.UnexpectedCodePathException;
 
-public abstract class AbstractApexNode<T extends AstNode> extends AbstractNode implements ApexNode<T> {
+public abstract class AbstractApexNode<T extends AstNode> extends AbstractApexNodeBase implements ApexNode<T> {
 
     protected final T node;
 
@@ -49,70 +47,6 @@ public abstract class AbstractApexNode<T extends AstNode> extends AbstractNode i
     }
 
     @Override
-    public int getBeginLine() {
-        if (this.beginLine > 0) {
-            return this.beginLine;
-        }
-        Node parent = jjtGetParent();
-        if (parent != null) {
-            return parent.getBeginLine();
-        }
-        throw new RuntimeException("Unable to determine beginning line of Node.");
-    }
-
-    @Override
-    public int getBeginColumn() {
-        if (this.beginColumn > 0) {
-            return this.beginColumn;
-        }
-        Node parent = jjtGetParent();
-        if (parent != null) {
-            return parent.getBeginColumn();
-        }
-        throw new RuntimeException("Unable to determine beginning column of Node.");
-    }
-
-    @Override
-    public int getEndLine() {
-        if (this.endLine > 0) {
-            return this.endLine;
-        }
-        Node parent = jjtGetParent();
-        if (parent != null) {
-            return parent.getEndLine();
-        }
-        throw new RuntimeException("Unable to determine ending line of Node.");
-    }
-
-    @Override
-    public int getEndColumn() {
-        if (this.endColumn > 0) {
-            return this.endColumn;
-        }
-        Node parent = jjtGetParent();
-        if (parent != null) {
-            return parent.getEndColumn();
-        }
-        throw new RuntimeException("Unable to determine ending column of Node.");
-    }
-
-    /**
-     * Accept the visitor. *
-     */
-    @Override
-    public Object childrenAccept(ApexParserVisitor visitor, Object data) {
-        if (children != null) {
-            for (int i = 0; i < children.length; ++i) {
-                @SuppressWarnings("unchecked")
-                // we know that the children here are all ApexNodes
-                ApexNode<T> apexNode = (ApexNode<T>) children[i];
-                apexNode.jjtAccept(visitor, data);
-            }
-        }
-        return data;
-    }
-
-    @Override
     public T getNode() {
         return node;
     }
@@ -131,15 +65,6 @@ public abstract class AbstractApexNode<T extends AstNode> extends AbstractNode i
             return false;
         }
     }
-
-
-
-
-    @Override
-    public final String getXPathNodeName() {
-        return this.getClass().getSimpleName().replaceFirst("^AST", "");
-    }
-
 
     public String getLocation() {
         if (hasRealLoc()) {
