+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/documentation/ApexDocRule.java
@@ -12,21 +12,17 @@ import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.antlr.runtime.ANTLRStringStream;
-import org.antlr.runtime.Token;
-
 import net.sourceforge.pmd.lang.apex.ast.ASTAnnotation;
+import net.sourceforge.pmd.lang.apex.ast.ASTFormalComment;
 import net.sourceforge.pmd.lang.apex.ast.ASTMethod;
 import net.sourceforge.pmd.lang.apex.ast.ASTModifierNode;
 import net.sourceforge.pmd.lang.apex.ast.ASTProperty;
 import net.sourceforge.pmd.lang.apex.ast.ASTUserClass;
 import net.sourceforge.pmd.lang.apex.ast.ASTUserInterface;
 import net.sourceforge.pmd.lang.apex.ast.ApexNode;
-import net.sourceforge.pmd.lang.apex.ast.ApexRootNode;
 import net.sourceforge.pmd.lang.apex.rule.AbstractApexRule;
 
 import apex.jorje.data.Locations;
-import apex.jorje.parser.impl.ApexLexer;
 import apex.jorje.semantic.ast.member.Parameter;
 import apex.jorje.semantic.ast.modifier.ModifierGroup;
 
@@ -41,40 +37,19 @@ public class ApexDocRule extends AbstractApexRule {
     private static final String UNEXPECTED_RETURN_MESSAGE = "Unexpected ApexDoc @return";
     private static final String MISMATCHED_PARAM_MESSAGE = "Missing or mismatched ApexDoc @param";
 
-    private boolean inClass;
     private boolean inTestClass;
-    private String source;
-    private List<TokenLocation> tokenLocations;
 
     @Override
     public Object visit(ASTUserClass node, Object data) {
-        if (inClass) {
-            super.visit(node, data);
-        } else {
-            inClass = true;
-            inTestClass = false;
-            buildTokens(node);
-            super.visit(node, data);
-            inClass = false;
-        }
-
+        super.visit(node, data);
         handleClassOrInterface(node, data);
-
         return data;
     }
 
-
     @Override
     public Object visit(ASTUserInterface node, Object data) {
-        if (inClass) {
-            super.visit(node, data);
-        } else {
-            buildTokens(node);
-            super.visit(node, data);
-        }
-
+        super.visit(node, data);
         handleClassOrInterface(node, data);
-
         return data;
     }
 
@@ -138,25 +113,6 @@ public class ApexDocRule extends AbstractApexRule {
         return data;
     }
 
-    private void buildTokens(ApexRootNode<?> node) {
-        source = node.getSource();
-        ANTLRStringStream stream = new ANTLRStringStream(source);
-        ApexLexer lexer = new ApexLexer(stream);
-
-        tokenLocations = new ArrayList<>();
-        Integer startIndex = 0;
-        Token token = lexer.nextToken();
-        Integer endIndex = lexer.getCharIndex();
-        while (token.getType() != Token.EOF) {
-            if (token.getType() != ApexLexer.WS) {
-                tokenLocations.add(new TokenLocation(startIndex, token.getText()));
-            }
-            startIndex = endIndex;
-            token = lexer.nextToken();
-            endIndex = lexer.getCharIndex();
-        }
-    }
-
     private void handleClassOrInterface(ApexNode<?> node, Object data) {
         ApexDocComment comment = getApexDocComment(node);
         if (comment == null) {
@@ -186,52 +142,22 @@ public class ApexDocRule extends AbstractApexRule {
     }
 
     private ApexDocComment getApexDocComment(ApexNode<?> node) {
-        String token = getApexDocToken(getApexDocIndex(node));
-        if (token == null) {
-            return null;
-        }
+        ASTFormalComment comment = node.getFirstChildOfType(ASTFormalComment.class);
+        if (comment != null) {
+            String token = comment.getToken();
 
-        boolean hasDescription = DESCRIPTION_PATTERN.matcher(token).find();
-        boolean hasReturn = RETURN_PATTERN.matcher(token).find();
+            boolean hasDescription = DESCRIPTION_PATTERN.matcher(token).find();
+            boolean hasReturn = RETURN_PATTERN.matcher(token).find();
 
-        ArrayList<String> params = new ArrayList<>();
-        Matcher paramMatcher = PARAM_PATTERN.matcher(token);
-        while (paramMatcher.find()) {
-            params.add(paramMatcher.group(1));
-        }
-
-        return new ApexDocComment(hasDescription, hasReturn, params);
-    }
-
-    private int getApexDocIndex(ApexNode<?> node) {
-        ASTAnnotation annotation = node.getFirstDescendantOfType(ASTAnnotation.class);
-        ApexNode<?> firstNode = annotation == null ? node : annotation;
-        int index = firstNode.getNode().getLoc().getStartIndex();
-        return source.lastIndexOf('\n', index);
-    }
-
-    private String getApexDocToken(int index) {
-        TokenLocation last = null;
-        for (TokenLocation location : tokenLocations) {
-            if (location.index >= index) {
-                if (last != null && last.token.startsWith("/**")) {
-                    return last.token;
-                }
-                return null;
+            ArrayList<String> params = new ArrayList<>();
+            Matcher paramMatcher = PARAM_PATTERN.matcher(token);
+            while (paramMatcher.find()) {
+                params.add(paramMatcher.group(1));
             }
-            last = location;
+
+            return new ApexDocComment(hasDescription, hasReturn, params);
         }
         return null;
-    }
-
-    private class TokenLocation {
-        int index;
-        String token;
-
-        TokenLocation(int index, String token) {
-            this.index = index;
-            this.token = token;
-        }
     }
 
     private class ApexDocComment {
